[Polyam] Refactor composer dropdowns
Port 3564a1555336c54de087ef67ce3568eff1371438 to polyam-glitch
